Contacts Service

+++ MedRepUserService/src/main/java/com/leosal/dbutils/DBEntity.java
@@ -1,5 +1,5 @@
 package com.leosal.dbutils;
 
 public interface DBEntity {
-	public Long getId();
+	public Integer getId();
 }
